Separated the View and Model into separate files
